+++ src/App.tsx
@@ -1,4 +1,4 @@
-import HomePage from './pages/home-page'
+import HomePage from './pages/HomePage'
 
 const App: React.FC = () => {
   return <HomePage />
