+++ src/pages/HomePage/SearchResults/MovieCard.tsx
